% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppm_from_date.R
\name{ppm_from_date}
\alias{ppm_from_date}
\title{Get the CO2 ppm value closest to a given date}
\usage{
ppm_from_date(date)
}
\arguments{
\item{date}{A character string of the format "yyyy-mm-dd"}
}
\value{
A floating point value, the closest data point to the input date
}
\description{
Given a date, retrieves the atmospheric CO2 concentration value the closest
to that date, using the data from the Mauna Loa observatory
}
\examples{
ppm_from_date("1978-06-05")

}
\references{
NOAA's \href{https://gml.noaa.gov/ccgg/trends/data.html}{Global Monitoring Laboratory data hub}.
}
